Better organize section helper file
